// frontend/src/interfaces/SectionProps.ts
interface SectionProps {
  width: string;
  height: string;
  minHeight: string;
  minWidth: string;
  contentBox?: boolean;
  block?: boolean;
  column?: boolean;
  justifyContent?: string;
  alignItems?: string;
  gap?: string;
  zIndex?: string;
  padding?: string;
  margin?: string;
  gridArea?: string;
  transition?: string;
}

export default SectionProps;